test(vsock-proto): group shared control bounds tests
