+++ crates/vsock-proto/src/payloads/exec_operation/tests/exec_control.rs
@@ -113,60 +113,6 @@ fn exec_control_roundtrip_and_rejects_malformed_payloads() {
     ));
 }
 #[test]
-fn exec_control_rejects_too_long_message_id() {
-    let message_id = "x".repeat(u16::MAX as usize + 1);
-
-    let err = encode_exec_control(7, NONCE, &message_id, b"body", 5000).unwrap_err();
-    assert!(matches!(
-        err,
-        ProtocolError::PayloadTooLarge("message_id", size) if size == message_id.len()
-    ));
-
-    let err = encode_exec_control_result(7, NONCE, &message_id, ExecControlStatus::Delivered, "")
-        .unwrap_err();
-    assert!(matches!(
-        err,
-        ProtocolError::PayloadTooLarge("message_id", size) if size == message_id.len()
-    ));
-
-    let diagnostic = "x".repeat(u16::MAX as usize + 1);
-    let err = encode_exec_control_result(
-        7,
-        NONCE,
-        "message",
-        ExecControlStatus::Delivered,
-        &diagnostic,
-    )
-    .unwrap_err();
-    assert!(matches!(
-        err,
-        ProtocolError::PayloadTooLarge("diagnostic", size) if size == diagnostic.len()
-    ));
-}
-#[test]
-fn exec_control_accepts_max_len_message_id_and_diagnostic() {
-    let max_message_id = "m".repeat(u16::MAX as usize);
-
-    let payload = encode_exec_control(7, NONCE, &max_message_id, b"body", 5000).unwrap();
-    assert_eq!(
-        decode_exec_control(&payload).unwrap().message_id,
-        max_message_id
-    );
-
-    let max_diagnostic = "d".repeat(u16::MAX as usize);
-    let result_payload = encode_exec_control_result(
-        7,
-        NONCE,
-        &max_message_id,
-        ExecControlStatus::Delivered,
-        &max_diagnostic,
-    )
-    .unwrap();
-    let decoded = decode_exec_control_result(&result_payload).unwrap();
-    assert_eq!(decoded.message_id, max_message_id);
-    assert_eq!(decoded.diagnostic, max_diagnostic);
-}
-#[test]
 fn exec_control_rejects_truncated_fields() {
     let payload = exec_control_payload();
     let request_timeout_offset = request_timeout_field_offset();
